--- tests/puppeteer/campaigns/functional/BO/13_shopParameters/03_productSettings/01_productsGeneral/06_enableDisableDefaultActivationStatus.js
@@ -1,101 +1,0 @@
-require('module-alias/register');
-const testContext = require('@utils/testContext');
-
-const baseContext = 'functional_BO_productSettings_enableDisableDefaultActivationStatus';
-// Using chai
-const {expect} = require('chai');
-const helper = require('@utils/helpers');
-const loginCommon = require('@commonTests/loginBO');
-// Importing pages
-const BOBasePage = require('@pages/BO/BObasePage');
-const LoginPage = require('@pages/BO/login');
-const DashboardPage = require('@pages/BO/dashboard');
-const ProductSettingsPage = require('@pages/BO/shopParameters/productSettings');
-const ProductsPage = require('@pages/BO/catalog/products');
-const AddProductPage = require('@pages/BO/catalog/products/add');
-// Importing data
-
-let browser;
-let page;
-
-// Init objects needed
-const init = async function () {
-  return {
-    boBasePage: new BOBasePage(page),
-    loginPage: new LoginPage(page),
-    dashboardPage: new DashboardPage(page),
-    productSettingsPage: new ProductSettingsPage(page),
-    productsPage: new ProductsPage(page),
-    addProductPage: new AddProductPage(page),
-  };
-};
-
-/*
-Enable default activation status
-Check that a new product is online by default
-Disable default activation status
- */
-describe('Enable/Disable default activation status', async () => {
-  // before and after functions
-  before(async function () {
-    browser = await helper.createBrowser();
-    page = await helper.newTab(browser);
-    this.pageObjects = await init();
-  });
-  after(async () => {
-    await helper.closeBrowser(browser);
-  });
-
-  // Login into BO and go to product settings page
-  loginCommon.loginBO();
-
-  it('should go to \'Shop parameters > Product Settings\' page', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'goToProductSettingsPageToEnableStatus', baseContext);
-    await this.pageObjects.boBasePage.goToSubMenu(
-      this.pageObjects.boBasePage.shopParametersParentLink,
-      this.pageObjects.boBasePage.productSettingsLink,
-    );
-    await this.pageObjects.boBasePage.closeSfToolBar();
-    const pageTitle = await this.pageObjects.productSettingsPage.getPageTitle();
-    await expect(pageTitle).to.contains(this.pageObjects.productSettingsPage.pageTitle);
-  });
-
-  it('should enable default activation status', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'enableDefaultActivationStatus', baseContext);
-    const result = await this.pageObjects.productSettingsPage.setDefaultActivationStatus(true);
-    await expect(result).to.contains(this.pageObjects.productSettingsPage.successfulUpdateMessage);
-  });
-
-  it('should go to \'Catalog > Products\' page', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'goToProductsPage', baseContext);
-    await this.pageObjects.boBasePage.goToSubMenu(
-      this.pageObjects.boBasePage.catalogParentLink,
-      this.pageObjects.boBasePage.productsLink,
-    );
-    const pageTitle = await this.pageObjects.productsPage.getPageTitle();
-    await expect(pageTitle).to.contains(this.pageObjects.productsPage.pageTitle);
-  });
-
-  it('should go to create product page and check that the new product is online by default', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'goToAddProductPageToCheckStatus', baseContext);
-    await this.pageObjects.productsPage.goToAddProductPage();
-    const online = await this.pageObjects.addProductPage.getOnlineButtonStatus();
-    await expect(online).to.be.true;
-  });
-
-  it('should go to \'Shop parameters > Product Settings\' page', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'goToProductSettingsPageToDisableStatus', baseContext);
-    await this.pageObjects.boBasePage.goToSubMenu(
-      this.pageObjects.boBasePage.shopParametersParentLink,
-      this.pageObjects.boBasePage.productSettingsLink,
-    );
-    const pageTitle = await this.pageObjects.productSettingsPage.getPageTitle();
-    await expect(pageTitle).to.contains(this.pageObjects.productSettingsPage.pageTitle);
-  });
-
-  it('should disable default activation status', async function () {
-    await testContext.addContextItem(this, 'testIdentifier', 'disableDefaultActivationStatus', baseContext);
-    const result = await this.pageObjects.productSettingsPage.setDefaultActivationStatus(false);
-    await expect(result).to.contains(this.pageObjects.productSettingsPage.successfulUpdateMessage);
-  });
-});
